LPS-116972 Moves code inside
